requirements-local.txt
boto3==1.26.124
botocore==1.29.124
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
jmespath==1.0.1
numpy==1.24.3
pandas==2.0.1
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
s3transfer==0.6.0
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
